New times:
```
1> perftest:test([ fun day7:solve/0 ], timer:seconds(5)).
Method                            Loops               Min               Max            Median           Average
empty loop    938554 *  500 = 469277000           0.002us           7.798us           0.002us           0.003us
day7:solve()     270 *   50 =     13500         306.620us         858.040us         351.580us         370.044us
```